Broke out sendTracklistPostRequest function to be imported to main.js. #2
